fix some of sonar issues
